Root cause: MakarovPM.cs and SilencedPistol.cs used `bulletNode as Bullet`
which incorrectly succeeds even for GDScript nodes in Godot's type system.
This caused pistol bullets to skip GDScript initialization, resulting in
broken homing functionality.

Changes:
- Use `bulletNode is Bullet` pattern instead of `as Bullet` for proper
  type detection of C# vs GDScript bullets
- Change shooter_id cast from (int) to (long) to avoid truncation of
  64-bit instance IDs (GDScript int is 64-bit)
- Apply same fix to BaseWeapon.cs for consistency

This also merges main to resolve conflicts from Issue #740 breaker fix.

+++ Scripts/AbstractClasses/BaseWeapon.cs
@@ -421,7 +421,7 @@ public abstract partial class BaseWeapon : Node2D
                 direction,
                 WeaponData?.BulletSpeed ?? 2500.0f,
                 WeaponData?.Damage ?? 1.0f,
-                (int)shooterId,
+                (long)shooterId,
                 GlobalPosition,
                 0.0f); // No stun duration in base weapon
         }
@@ -436,7 +436,7 @@ public abstract partial class BaseWeapon : Node2D
             }
             if (owner != null)
             {
-                bullet.Set("shooter_id", (int)shooterId);
+                bullet.Set("shooter_id", (long)shooterId);
             }
             bullet.Set("shooter_position", GlobalPosition);
         }
